hyprland/workspaces: break up headers
